refactor: refactor for tests
