+++ controllers/contentController.js
@@ -1,14 +1,14 @@
-const { gitFileContent, gitFileTree } = require('../utils/git');
+const { Git } = require('../utils/git');
 const { buildBreadcrumbs } = require('../utils/navigation');
 
 module.exports = function(req, res, next) {
   const { hash } = req.params;
   const path = req.params[0].split('/').filter(Boolean);
 
-  gitFileTree(hash, path.join('/'))
+  Git.gitFileTree(hash, path.join('/'))
     .then(function([file]) {
       if (file && file.type === 'blob') {
-        return gitFileContent(file.hash);
+        return Git.gitFileContent(file.hash);
       }
     })
     .then(
