+++ controllers/indexController.js
@@ -1,8 +1,8 @@
-const { gitHistory } = require('../utils/git');
+const { Git } = require('../utils/git');
 const { buildFolderUrl, buildBreadcrumbs } = require('../utils/navigation');
 
 module.exports = function(req, res) {
-  gitHistory(1, 20).then(
+  Git.gitHistory(1, 20).then(
     history => {
       const list = history.map(item => ({
         ...item,
