+++ utils/helpers.js
@@ -1,0 +1,19 @@
+function parseFileTreeItem(line) {
+  const [info, path] = line.split('\t');
+  const [, type, hash] = info.split(' ');
+
+  return { type, hash, path };
+}
+
+function parseHistoryItem(line) {
+  const [hash, author, timestamp, msg] = line.split('\t');
+
+  return {
+    hash,
+    author,
+    timestamp,
+    msg
+  };
+}
+
+module.exports = { parseFileTreeItem, parseHistoryItem };
